Materializer updated
