+++ src/zenml/integrations/huggingface/materializers/huggingface_pt_model_materializer.py
@@ -1,0 +1,60 @@
+#  Copyright (c) ZenML GmbH 2021. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+import importlib
+import os
+from tempfile import TemporaryDirectory
+from typing import Any, Type
+
+from transformers import AutoConfig, PreTrainedModel
+
+from zenml.artifacts import ModelArtifact
+from zenml.io import fileio
+from zenml.materializers.base_materializer import BaseMaterializer
+
+DEFAULT_PT_MODEL_DIR = "hf_pt_model"
+
+
+class HFPTModelMaterializer(BaseMaterializer):
+    """Materializer to read torch model to and from huggingface pretrained model."""
+
+    ASSOCIATED_TYPES = (PreTrainedModel,)
+    ASSOCIATED_ARTIFACT_TYPES = (ModelArtifact,)
+
+    def handle_input(self, data_type: Type[Any]) -> PreTrainedModel:
+        """Reads HFModel"""
+        super().handle_input(data_type)
+
+        config = AutoConfig.from_pretrained(
+            os.path.join(self.artifact.uri, DEFAULT_PT_MODEL_DIR)
+        )
+        architecture = config.architectures[0]
+        model_cls = getattr(
+            importlib.import_module("transformers"), architecture
+        )
+        return model_cls.from_pretrained(
+            os.path.join(self.artifact.uri, DEFAULT_PT_MODEL_DIR)
+        )
+
+    def handle_return(self, model: Type[Any]) -> None:
+        """Writes a Model to the specified dir.
+        Args:
+            PreTrainedModel: The Torch Model to write.
+        """
+        super().handle_return(model)
+        temp_dir = TemporaryDirectory()
+        model.save_pretrained(temp_dir.name)
+        fileio.copy_dir(
+            temp_dir.name, os.path.join(self.artifact.uri, DEFAULT_PT_MODEL_DIR)
+        )
